pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gravia"
version = "1.0.0"
description = "Professional research factory: paper analysis, manuscript synthesis, interactive presentations."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Hamed Nejat", email = "hamed.nejat@vanderbilt.edu"},
]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "jinja2>=3.1",
    "markdown>=3.5",
    "pymdown-extensions>=10.0",
    "PyMuPDF>=1.24",
    "pypandoc>=1.13",
    "Pillow>=10.0",
    "plotly>=5.18",
    "PyYAML>=6.0",
    "rich>=13.0",
    "tomli>=2.0;python_version<'3.11'",
    "mlx-lm>=0.10.0",
    "cryptography>=42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
pdf-export = [
    "playwright>=1.40",
]

[project.scripts]
gravia = "gravia.cli:main"

[tool.setuptools.packages.find]
include = ["gravia*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
